Move converters to separate packages and rename old converters
Jira: MODKBEKBJ-109
Renamed old converters to *RequestConverter, because they only have
methods for converting requests now
