+++ src/main/java/org/folio/rest/converter/packages/PackageCollectionConverter.java
@@ -1,4 +1,4 @@
-package org.folio.rest.converter;
+package org.folio.rest.converter.packages;
 
 import java.util.List;
 import java.util.stream.Collectors;
