+++ src/main/java/org/folio/rest/converter/providers/ProviderConverter.java
@@ -1,12 +1,13 @@
-package org.folio.rest.converter;
+package org.folio.rest.converter.providers;
 
-import static org.folio.rest.converter.VendorConverter.createEmptyProviderRelationships;
+import static org.folio.rest.converter.providers.ProviderRequestConverter.createEmptyProviderRelationships;
 import static org.folio.rest.util.RestConstants.PACKAGES_TYPE;
 import static org.folio.rest.util.RestConstants.PROVIDERS_TYPE;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.folio.rest.converter.util.CommonAttributesConverter;
 import org.folio.rest.jaxrs.model.MetaDataIncluded;
 import org.folio.rest.jaxrs.model.PackageCollection;
 import org.folio.rest.jaxrs.model.Packages;
