+++ src/main/java/org/folio/rest/converter/util/CommonResourceConverter.java
@@ -1,4 +1,4 @@
-package org.folio.rest.converter;
+package org.folio.rest.converter.util;
 
 import org.folio.rest.jaxrs.model.ResourceDataAttributes;
 import org.folio.rmapi.model.CustomerResources;
